Move Schedulers and Aggregators into folders
