+++ lib/broken_record/scanner.rb
@@ -1,15 +1,15 @@
 require 'broken_record/class_finder'
 require 'broken_record/job'
-require 'broken_record/job_scheduler'
-require 'broken_record/parallel_job_scheduler'
-require 'broken_record/external_job_scheduler'
-require 'broken_record/result_aggregator'
-require 'broken_record/slack_aggregator'
-require 'broken_record/datadog_aggregator'
-require 'broken_record/console_aggregator'
-require 'broken_record/json_aggregator'
-require 'broken_record/multi_aggregator'
-require 'broken_record/bugsnag_aggregator'
+require 'broken_record/schedulers/job_scheduler'
+require 'broken_record/schedulers/parallel_job_scheduler'
+require 'broken_record/schedulers/external_job_scheduler'
+require 'broken_record/aggregators/result_aggregator'
+require 'broken_record/aggregators/slack_aggregator'
+require 'broken_record/aggregators/datadog_aggregator'
+require 'broken_record/aggregators/console_aggregator'
+require 'broken_record/aggregators/json_aggregator'
+require 'broken_record/aggregators/multi_aggregator'
+require 'broken_record/aggregators/bugsnag_aggregator'
 
 module BrokenRecord
   class Scanner
@@ -20,7 +20,7 @@ module BrokenRecord
         BrokenRecord::Config.before_scan_callbacks.each { |callback| callback.call }
 
         scheduler_class = BrokenRecord::Config.job_scheduler_class.constantize
-        raise "Invalid job scheduler" unless scheduler_class.ancestors.include?(BrokenRecord::JobScheduler)
+        raise "Invalid job scheduler" unless scheduler_class.ancestors.include?(BrokenRecord::Schedulers::JobScheduler)
         scheduler = scheduler_class.new(classes, aggregator, BrokenRecord::Config.job_scheduler_options)
         scheduler.run
       end
